libm: fix narrowing-FMA, sinh/cosh overflow gap, totalorder/setpayload long double (MATH-03/04/05/06)
MATH-05 (math_narrowing.c): ffma() rounded each double operand to float BEFORE
the multiply, defeating the single-rounding contract of a narrowing FMA.  Now
promotes to long double and rounds once: (float)fmal(x,y,z), matching ffmal.

MATH-06 (math_trig.c): sinh/cosh called exp(x) directly and thus inherited
exp's overflow at 709.78, spuriously returning INFINITY+ERANGE for inputs in
(709.78, 710.47] whose true magnitude is still finite (e^|x|/2 < DBL_MAX until
|x| > ln(2)+ln(DBL_MAX)=710.4758).  Added a large-arg branch computing e^|x|/2
via the split (0.5*e^{|x|/2})*e^{|x|/2}; only |x|>710.4758 sets ERANGE.

MATH-03 (math_totalorder.c): totalorderl/totalordermagl collapsed long double
to double, dropping the 64-bit significand to 53 bits so distinct 80-bit values
compared order-equal (breaks antisymmetry).  Now decode the 80-bit x87 encoding
(sign/15-bit exp/64-bit significand) directly.

MATH-04 (math_totalorder.c): setpayloadl forced byte-9 bit7 (the SIGN) on, so
every result was a negative pseudo-NaN, and both setpayloadl/setpayloadsigl
smeared the payload into non-payload mantissa bytes.  Set the integer+quiet bits
in byte 7 instead, sign follows x, payload stays in the low 32 bits.

libm.a/.so rebuild clean under -Werror; host harness 33/33 numerical checks pass.

+++ lib/m/src/math_narrowing.c
@@ -179,8 +179,11 @@ double ddivl(long double a, long double b) {
  * ============================================================ */
 
 double ffma(double x, double y, double z) {
-    /* All-float result: fmaf rounds x*y+z once to float. */
-    return fmaf((float)x, (float)y, (float)z);
+    /* float result, double operands: fuse in long double (which holds
+     * every double operand exactly), then round once to float.  Casting
+     * the operands to float first (fmaf) would discard ~29 bits of each
+     * operand before the multiply, violating the C23 narrowing contract. */
+    return (float)fmal(x, y, z);
 }
 
 double ffmal(long double x, long double y, long double z) {
